+++ src/flow/vector.hpp
@@ -1,5 +1,5 @@
 namespace flow {
 	class Vector : public Point {
-		Vector(int x=0, int y=0, int z=0) : Point(x, y, z) {};
+		Vector(double x=0, double y=0, double z=0) : Point(x, y, z) {};
 	};
 }
